坐标、旋转和缩放全部改成双精度浮点，FRotator和FTransform改成了模板。在处理坐标系的时候，原来是在base pass将横坐标反转，来在opengl下实现和ue相同的坐标系，现在改成在最后render to screen的时候翻转uv

+++ Assets/Shaders/GlobalShaders/LightShaderParameters.glsl
@@ -24,7 +24,7 @@ struct FLightShaderParameters
     int     ShadowMappingStartIndex;
     int     ShadowMappingEndIndex;
     vec4    lightColor;
-    vec3    lightPosition;
+    dvec3   lightPosition;
     vec3    lightDirection;
     int     lightType;
     float   lightIntensity;
